% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggslayer.R
\name{addContinuousDGGSLayer}
\alias{addContinuousDGGSLayer}
\title{Title}
\usage{
addContinuousDGGSLayer(
  map,
  layer,
  tid = NULL,
  filter = NULL,
  layerId = NULL,
  group = NULL,
  options = dggsContinuousTileOptions()
)
}
\arguments{
\item{options}{}
}
\value{

}
\description{
Title
}
