% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggslayer.R
\name{addNominalDGGSLayer}
\alias{addNominalDGGSLayer}
\title{Adds a DGGS Nominal Type layer.}
\usage{
addNominalDGGSLayer(
  map,
  layer,
  tid = NULL,
  filter = NULL,
  layerId = NULL,
  group = NULL,
  options = dggsNominalTileOptions()
)
}
\arguments{
\item{map}{a leaflet map}

\item{layer}{the dataset name. it is the table name which has a dggid,key,tid,value structure in nz}

\item{tid}{tid value.}

\item{filter}{filter data, can be in a form of key='something', it is mostly an SQL where statement}

\item{options}{use dggsNominalTileOptions() function, the mandetory option for this type of the layer is classNames which is a 3dimentional list in this format c("Lable",value,"HexColor")}
}
\value{

}
\description{
Adds a DGGS Nominal Type layer.
}
\examples{
m <- leaflet() \%>\%
}
