% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggslayer.R
\name{addNumericalDGGSLayer}
\alias{addNumericalDGGSLayer}
\title{Title}
\usage{
addNumericalDGGSLayer(
  map,
  layer,
  tid = NULL,
  filter = NULL,
  layerId = NULL,
  group = NULL,
  options = dggsNumericalTileOptions()
)
}
\arguments{
\item{options}{}
}
\value{

}
\description{
Title
}
