% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggslayer.R
\name{setServerAddress}
\alias{setServerAddress}
\title{setServerAddress
Configure Tileserves's Access URL.}
\usage{
setServerAddress(map, server_URL)
}
\arguments{
\item{server_URL}{}
}
\value{

}
\description{
setServerAddress
Configure Tileserves's Access URL.
}
\examples{
m <- leaflet() \%>\%
setServerAddress("https://spatial.wlu.ca/tileserver/")\%>\%
}
