% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggslayer.R
\name{setServerAddress}
\alias{setServerAddress}
\title{setServerAddress}
\usage{
setServerAddress(map, server_URL)
}
\arguments{
\item{map}{a leaflet map}

\item{server_URL}{URL of the server. This must be set if you are connecting to another server than IDEAS tile server}
}
\value{
leaflerR map object
}
\description{
Configure Tileserves's Access URL.
}
\examples{
\dontrun{ m <- leaflet() \%>\%
setServerAddress("https://spatial.wlu.ca/tileserver/")
}
}
